Explicitly check the length field at both RX and TX.
IEEE 802.15.4 allows data packets of 9-127 bytes and acknowledgements of length 5.
This patch adds a check for the received packet length and rejects non-compliant packets.

For transmitted packets only 7 bits are considered as packet length and the
MSB of the byte holding the length field is reset to zero (to comply with the standard).